factor out opt key hold tracking into useOptHeld
